+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v3.0.0
+
+- Switch to random_pet
+- Add hostnames variable to give full control over the gateway endpoints
+- Remove Konga
+
 ## v2.3.3
 
 - Use id_bg